[iOS] Further move WebKit navigation & download logic into shared code
This code also modifies the response policy decision handler to match Chromium, such that any request that isn't meant to render in the web view will now respond with a `download` policy. This change means that prior handling of calendar invites & mobile configurations are moved into download logic (which also matches Chromium).

This change also removes dead `HTTPDownload` logic and its associated helper (`DownloadHelper`)
