+++ ios/brave-ios/Sources/Brave/Frontend/Browser/BrowserViewController/BVC+TabWebDelegate.swift
@@ -5,6 +5,7 @@
 
 import BraveCore
 import BraveShared
+import Data
 import Foundation
 import Preferences
 import Shared
@@ -43,6 +44,13 @@ protocol TabWebDelegate: AnyObject {
     proposedCredential credential: URLCredential?,
     previousFailureCount: Int
   ) async -> URLCredential?
+  func tab(
+    _ tab: Tab,
+    createNewTabWithRequest request: URLRequest,
+    configuration: WKWebViewConfiguration
+  ) -> Tab?
+  func tab(_ tab: Tab, shouldBlockJavaScriptForRequest request: URLRequest) -> Bool
+  func tab(_ tab: Tab, shouldBlockUniversalLinksForRequest request: URLRequest) -> Bool
 }
 
 /// Media device capture types that a web page may request
@@ -102,6 +110,14 @@ extension TabWebDelegate {
     previousFailureCount: Int
   ) async -> URLCredential? {
     return nil
+  }
+
+  func tab(_ tab: Tab, shouldBlockJavaScriptForRequest request: URLRequest) -> Bool {
+    return false
+  }
+
+  func tab(_ tab: Tab, shouldBlockUniversalLinksForRequest request: URLRequest) -> Bool {
+    return false
   }
 }
 
@@ -402,6 +418,31 @@ extension BrowserViewController: TabWebDelegate {
       }
     }
   }
+
+  func tab(_ tab: Tab, shouldBlockJavaScriptForRequest request: URLRequest) -> Bool {
+    guard let documentTargetURL = request.mainDocumentURL else { return false }
+    let domainForShields = Domain.getOrCreate(
+      forUrl: documentTargetURL,
+      persistent: !tab.isPrivate
+    )
+    return domainForShields.isShieldExpected(.noScript, considerAllShieldsOption: true)
+  }
+
+  func tab(_ tab: Tab, shouldBlockUniversalLinksForRequest request: URLRequest) -> Bool {
+    func isYouTubeLoad() -> Bool {
+      guard let domain = request.mainDocumentURL?.baseDomain else {
+        return false
+      }
+      let domainsWithUniversalLinks: Set<String> = ["youtube.com", "youtu.be"]
+      return domainsWithUniversalLinks.contains(domain)
+    }
+    if tab.isPrivate || !Preferences.General.followUniversalLinks.value
+      || (Preferences.General.keepYouTubeInBrave.value && isYouTubeLoad())
+    {
+      return true
+    }
+    return false
+  }
 }
 
 extension BrowserViewController {
@@ -571,5 +612,50 @@ extension BrowserViewController {
     } catch {
       return nil
     }
+  }
+
+  func tab(
+    _ tab: Tab,
+    createNewTabWithRequest request: URLRequest,
+    configuration: WKWebViewConfiguration
+  ) -> Tab? {
+    guard !request.isInternalUnprivileged,
+      let navigationURL = request.url,
+      navigationURL.shouldRequestBeOpenedAsPopup()
+    else {
+      print("Denying popup from request: \(request)")
+      return nil
+    }
+
+    if let currentTab = tabManager.selectedTab {
+      screenshotHelper.takeScreenshot(currentTab)
+    }
+
+    // If the page uses `window.open()` or `[target="_blank"]`, open the page in a new tab.
+    // IMPORTANT!!: WebKit will perform the `URLRequest` automatically!! Attempting to do
+    // the request here manually leads to incorrect results!!
+    let newTab = tabManager.addPopupForParentTab(tab, configuration: configuration)
+
+    newTab.setVirtualURL(URL(string: "about:blank"))
+
+    toolbarVisibilityViewModel.toolbarState = .expanded
+
+    // Wait until WebKit starts the request before selecting the new tab, otherwise the tab manager may
+    // restore it as if it was a dead tab.
+    var observation: NSKeyValueObservation?
+    observation = newTab.webView?.observe(
+      \.url,
+      changeHandler: { [weak self, weak newTab] webView, _ in
+        _ = observation  // Silence write but not read warning
+        observation = nil
+        guard let self = self, let tab = newTab else { return }
+
+        // When a child tab is being selected, dismiss any popups on the parent tab
+        tab.parent?.shownPromptAlert?.dismiss(animated: false)
+        self.tabManager.selectTab(tab)
+      }
+    )
+
+    return newTab
   }
 }
